requirements/dev.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
backports.zoneinfo==0.2.1
billiard==4.2.1
boto3==1.37.10
botocore==1.37.10
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.20
django-autoslug==1.9.9
django-cors-headers==4.4.0
django-filter==24.3
django-ses==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
importlib_resources==6.4.5
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.4.2
pkgutil_resolve_name==1.3.10
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
referencing==0.35.1
rpds-py==0.20.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
